encoders and decoders inheriting from Module adding encoders_weights dict
