+++ segmentation_models_pytorch/base/model.py
@@ -1,22 +1,7 @@
 import torch
-from torch import nn
+from .module import Module
 from . import initialization as init
 from ..layers import out_modules
-
-
-class Module(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    def save(self, path='new_state_dict.pth'):
-        torch.save(self.state_dict(), path)
-        print('saving module in', path)
-
-    def load(self, path=None):
-        if not path:
-            path = 'new_state_dict.pth'
-        self.load_state_dict(torch.load(path))
-        print('loading module from', path)
 
 
 class Model(Module):
