+++ segmentation_models_pytorch/encoders/inceptionv4.py
@@ -28,9 +28,10 @@ from pretrainedmodels.models.inceptionv4 import InceptionV4, BasicConv2d
 from pretrainedmodels.models.inceptionv4 import pretrained_settings
 
 from ._base import EncoderMixin
+from ..base.module import Module
 
 
-class InceptionV4Encoder(InceptionV4, EncoderMixin):
+class InceptionV4Encoder(InceptionV4, EncoderMixin, Module):
     def __init__(self, stage_idxs, out_channels, depth=5, **kwargs):
         super().__init__(**kwargs)
         self._stage_idxs = stage_idxs
